moves RelateOnCopyHandler to Core and changes from ApplicationStarting to ApplicationStarted
